src/test/java/dev/kmunton/year2024/day5/Day5Test.java
package dev.kmunton.year2024.day5;

import static org.junit.jupiter.api.Assertions.assertEquals;

import java.util.List;
import org.junit.jupiter.api.Test;

class Day5Test {

  private static Day5 day;

  @Test
  void part1_success() {
    List<String> input = """
        47|53
        97|13
        97|61
        97|47
        75|29
        61|13
        75|53
        29|13
        97|29
        53|29
        61|53
        97|53
        61|29
        47|13
        75|47
        97|75
        47|61
        75|61
        47|29
        75|13
        53|13
                
        75,47,61,53,29
        97,61,53,29,13
        75,29,13
        75,97,47,61,53
        61,13,29
        97,13,75,29,47""".lines().toList();
    day = new Day5(input);
    assertEquals(143L, day.part1());
  }

  @Test
  void part2_success() {
    List<String> input = """
        47|53
        97|13
        97|61
        97|47
        75|29
        61|13
        75|53
        29|13
        97|29
        53|29
        61|53
        97|53
        61|29
        47|13
        75|47
        97|75
        47|61
        75|61
        47|29
        75|13
        53|13
                
        75,47,61,53,29
        97,61,53,29,13
        75,29,13
        75,97,47,61,53
        61,13,29
        97,13,75,29,47""".lines().toList();
    day = new Day5(input);
    assertEquals(123L, day.part2());
  }

}
